requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
cryptography==2.3.1
flake8==3.5.0
future==0.16.0
idna==2.7
mccabe==0.6.1
pycodestyle==2.3.1
pycparser==2.19
pyflakes==1.6.0
python-telegram-bot==11.1.0
six==1.11.0
